- implemented logic - added exchangePointers() to flexible_array

+++ src/gui_debug_mouse.hpp
@@ -108,9 +108,9 @@ void GuiDebugMouse(gamestate_t & gamestate, GuiDebugMouseState* state)
     if (state->WindowBox000Active)
     {
         static char mou_x_buffer[10];
-        _itoa_s(gamestate.mousePosition.x, mou_x_buffer, 10, 10); // Cast MousePosition.x integer to c_string
+        _itoa_s((int)gamestate.mousePosition.x, mou_x_buffer, 10, 10); // Cast MousePosition.x integer to c_string
         static char mou_y_buffer[10];
-        _itoa_s(gamestate.mousePosition.y, mou_y_buffer, 10, 10); // Cast MousePosition.y integer to c_string
+        _itoa_s((int)gamestate.mousePosition.y, mou_y_buffer, 10, 10); // Cast MousePosition.y integer to c_string
 
         static char mouse_lval_str[6];
         if (gamestate.mouse_L) {
